Fixed Issue BSAPP-489
Implemented a mechanism, that automatically sets the higher year as the min year and the lower year as the max year of a competition class.

Updated Tests accordingly.

Fixed an issue in the business logic when only having one year as min and max.

Refactored duplicate code in this class and extracted it into two methods.